requirements.txt
instaloader>=4.10
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0
